[Core] Provide a unique id of the current scenario to the hooks.
Provide methods to access the pickle uri and the pickle lines, which
uniquely identify the current scenario. In addition also put back the
getId() method. Even though it is not possible to provide the same id
string as when using Gherkin v2.12.2, hooks using Scenario.getId()
will still compile and still get a unique id for the scenario through
that method.